# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  ready: 'ready',
  analyzing: 'analyzing',
  draft: 'draft',
} as const;

export type RiskSeverity = typeof RiskSeverity[keyof typeof RiskSeverity];


export const RiskSeverity = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Risk {
  severity: RiskSeverity;
  title: string;
  explanation: string;
  /** @nullable */
  excerpt: string | null;
}

export interface Document {
  id: string;
  title: string;
  /** @nullable */
  fileName: string | null;
  status: DocumentStatus;
  sourceText: string;
  /** @nullable */
  summary: string | null;
  risks: Risk[];
  riskCount: number;
  /** @nullable */
  analyzedAt: string | null;
  createdAt: string;
}

export interface DocumentInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  fileName?: string | null;
  /** @minLength 1 */
  sourceText: string;
}

export interface AnalysisInput {
  sourceText?: string;
  model?: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: string;
  documentId: string;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface ChatInput {
  /** @minLength 1 */
  question: string;
  model?: string;
}

export interface Overview {
  documentCount: number;
  analyzedCount: number;
  openRiskCount: number;
  /** @nullable */
  lastAnalyzedAt: string | null;
}